Refactor Bijoy and Unicode conversion implementations; move to dedicated index files and remove deprecated code
